refactor: Search에 관련된 메서드 위치 이동(Video->Search)
