+++ src/main/java/com/example/demo/src/search/SearchDao.java
@@ -14,4 +14,9 @@ public class SearchDao {
         String query = "insert into Search (keyword) VALUES (?)";
         return this.jdbcTemplate.update(query, keyword);
     }
+	public String getTheMostSearchedWord() {
+		String theMostSearchedWordQuery = "select keyword from Search group by keyword order by count(keyword) desc limit 1";
+		return this.jdbcTemplate.queryForObject(theMostSearchedWordQuery,
+			String.class);
+	}
 }
