+++ src/main/java/com/example/demo/src/video/VideoDao.java
@@ -312,12 +312,6 @@ public class VideoDao {
 		return videos;
 	}
 
-	public String getTheMostSearchedWord() {
-		String theMostSearchedWordQuery = "select keyword from Search group by keyword order by count(keyword) desc limit 1";
-		return this.jdbcTemplate.queryForObject(theMostSearchedWordQuery,
-			String.class);
-	}
-
 	public List<Video.getVideoResDto> getAwardVideosByGenre(int videoType, String genre) {
 		String isMovieOrSeries = videoType == 0 ? "=" : "<";
 		String query = "select V.* from GenreContact as GC "
